features2d(test): extract common extract/invariance test code
to share with opencv_contrib/xfeatures2d
